Complete sequence in build.rb ci_utils added
